requirements.txt
websocket-client==1.5.1
requests==2.28.2